prototype/frameworks/mcp/openapi/requirements.txt
anyio
click
httpx
pydantic
pyyaml
mcp
uvicorn
starlette